requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
attrs=22.1.0=py37hecd8cb5_0
blas=1.0=openblas
bottleneck=1.3.5=py37h67323c0_0
branca=0.6.0=pypi_0
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
bzip2=1.0.8=h1de35cc_0
c-ares=1.18.1=hca72f7f_0
ca-certificates=2022.10.11=hecd8cb5_0
cairo=1.16.0=h691a603_2
certifi=2022.12.7=py37hecd8cb5_0
cfitsio=3.470=hbd21bf8_7
charset-normalizer=2.1.1=pypi_0
click=7.1.2=pyhd3eb1b0_0
click-plugins=1.1.1=pyhd3eb1b0_0
cligj=0.7.2=py37hecd8cb5_0
curl=7.86.0=h6c40b1e_0
cycler=0.11.0=pyhd3eb1b0_0
eigen=3.3.7=haf03e11_1
expat=2.4.9=he9d5cce_0
fftw=3.3.9=h9ed2024_1
fiona=1.8.13.post1=py37h9c05f0f_0
folium=0.13.0=pypi_0
fontconfig=2.13.1=hb62acd4_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=hd8bbffd_0
freexl=1.0.6=h9ed2024_0
gdal=3.0.2=py37h339f5b9_4
geopandas=0.9.0=py_1
geopandas-base=0.9.0=py_1
geos=3.8.0=hb1e8313_0
geotiff=1.7.0=hf77b895_0
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
glib=2.69.1=h8346a28_1
graphite2=1.3.14=he9d5cce_1
gst-plugins-base=1.14.0=h4180768_2
gstreamer=1.14.0=h0fc69c2_2
harfbuzz=4.3.0=h3f129a2_0
hdf4=4.2.13=h39711bb_2
hdf5=1.10.6=h10fe05b_1
icu=58.2=h0a44026_3
idna=3.4=pypi_0
intel-openmp=2021.4.0=hecd8cb5_3538
jinja2=3.1.2=pypi_0
joblib=1.1.1=py37hecd8cb5_0
jpeg=9e=hca72f7f_0
json-c=0.16=hca72f7f_0
kealib=1.4.14=h7cdcae2_1
kiwisolver=1.4.2=py37he9d5cce_0
krb5=1.19.2=hcd88c3b_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libboost=1.73.0=h3fa6bed_12
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libclang=12.0.0=default_hbc2896b_2
libcurl=7.86.0=ha585b31_0
libcxx=14.0.6=h9765a3e_0
libdap4=3.19.1=h3d3e54a_0
libdeflate=1.8=h9ed2024_5
libedit=3.1.20210910=hca72f7f_0
libev=4.33=h9ed2024_1
libffi=3.3=hb1e8313_2
libgdal=3.0.2=h92a8043_4
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libiconv=1.16=hca72f7f_2
libkml=1.3.0=h952ee91_5
libllvm12=12.0.0=h9b2ccf5_3
libnetcdf=4.8.1=h24cb85c_1
libnghttp2=1.46.0=ha29bfda_0
libopenblas=0.3.21=h54e7dc3_0
libpng=1.6.37=ha441bb4_0
libpq=12.9=h1c9f633_3
libspatialindex=1.9.3=h23ab428_0
libspatialite=4.3.0a=h268bb56_20
libssh2=1.10.0=h0a4fc7d_0
libtiff=4.4.0=h2cd0358_2
libwebp=1.2.4=h56c3ce4_0
libwebp-base=1.2.4=hca72f7f_0
libxml2=2.9.14=hbf8cd5e_0
libxslt=1.1.35=h5b33f42_0
libzip=1.8.0=h272c8d6_0
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.3=h23ab428_1
mapclassify=2.4.3=pyhd3eb1b0_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.3=py37hecd8cb5_0
matplotlib-base=3.5.3=py37hfb0c5b7_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py37h9ed2024_0
munch=2.5.0=pyhd3eb1b0_0
munkres=1.1.4=py_0
ncurses=6.3=hca72f7f_3
networkx=2.6.3=pyhd3eb1b0_0
nspr=4.33=he9d5cce_0
nss=3.74=h47edf6a_0
numexpr=2.8.4=py37h57a7bef_0
numpy=1.21.5=py37h0f1bd0b_3
numpy-base=1.21.5=py37hbda7086_3
opencv=4.6.0=py37h0e6eb04_2
openjpeg=2.4.0=h66ea3da_0
openssl=1.1.1s=hca72f7f_0
overpy=0.6=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py37h743cdd8_0
pcre=8.45=h23ab428_0
pillow=9.2.0=py37hde71d04_1
pip=22.2.2=py37hecd8cb5_0
pixman=0.40.0=h9ed2024_1
poppler=0.81.0=h031bf0f_2
poppler-data=0.4.11=hecd8cb5_0
proj=6.2.1=hfd5b9e3_0
pyparsing=3.0.9=py37hecd8cb5_0
pyproj=2.6.1.post1=py37hdfdfadc_1
python=3.7.13=hdfd78df_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2022.7=py37hecd8cb5_0
qt-main=5.15.2=h719ae48_7
qt-webengine=5.15.9=h90a370e_4
qtwebkit=5.212=h24dc246_4
readline=8.2=hca72f7f_0
requests=2.28.1=pypi_0
rtree=0.9.7=py37hecd8cb5_1
scikit-learn=1.0.2=py37hae1ba45_1
scipy=1.7.3=py37h0ed0bf1_2
setuptools=65.5.0=py37hecd8cb5_0
shapely=1.8.4=py37ha3a5ccb_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.40.0=h880c91c_0
threadpoolctl=2.2.0=pyh0d69192_0
tiledb=2.3.3=h74f4c1e_2
tk=8.6.12=h5d9f67b_0
tornado=6.2=py37hca72f7f_0
typing_extensions=4.3.0=py37hecd8cb5_0
urllib3=1.26.13=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xerces-c=3.2.4=hd2b157f_0
xz=5.2.6=hca72f7f_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.2=hcb37349_0
